lab-big-data/requirements_pipeline.txt
tensorflow
keras
numpy
pillow
requests
apache-beam
apache-beam[gcp]